+++ spike/riscv/insns/ss_app_mod_siz_dec.h
@@ -2,7 +2,7 @@ auto &destReg = P.SU.registers[insn.uve_rd()];
 auto sizeReg = insn.uve_mod_size();
 auto dispReg = insn.uve_mod_disp();
 
-size_t size = READ_REG(sizeReg);
+unsigned int size = READ_REG(sizeReg);
 int32_t disp = READ_REG(dispReg);
 
 std::visit([&](auto &reg) {
